A rickety 'create exam' option is now available outside the admin form, which does allow us to query only the correct syllabus. Much work still to go though.

+++ tracker/models.py
@@ -141,7 +141,7 @@ class Exam(models.Model):
 class Question(models.Model):
     exam = models.ForeignKey(Exam, on_delete=models.CASCADE)
     qnumber = models.CharField(max_length=100)
-    qorder = models.IntegerField()
+    qorder = models.DecimalField(decimal_places=2, max_digits=6)
     syllabuspoint = models.ManyToManyField(SyllabusPoint)
     maxscore = models.IntegerField()
